Refactored dialog window to be a separate component
